+++ packages/@cosmos-building-blocks/pipeline/src/pipeline/cdk-pipeline.ts
@@ -1,42 +1,39 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 import { Construct, Stack, IConstruct, Lazy, IResolvable, PhysicalName } from '@aws-cdk/core';
-import { IRepository, Repository } from '@aws-cdk/aws-codecommit';
+import { IRepository } from '@aws-cdk/aws-codecommit';
 import { Pipeline, Artifact, IAction, IStage } from '@aws-cdk/aws-codepipeline';
-import {
-  CodeCommitSourceAction,
-  CodeBuildAction,
-  CodeCommitTrigger,
-  ManualApprovalAction,
-  CodeBuildActionProps,
-} from '@aws-cdk/aws-codepipeline-actions';
+import { CodeBuildAction, ManualApprovalAction, CodeBuildActionProps } from '@aws-cdk/aws-codepipeline-actions';
 import {
   Project,
   LinuxBuildImage,
   BuildEnvironmentVariable,
-  Source,
   Artifacts,
   BuildEnvironmentVariableType,
+  ComputeType,
 } from '@aws-cdk/aws-codebuild';
 import { IRole, Role, CompositePrincipal, ServicePrincipal } from '@aws-cdk/aws-iam';
 import { IVpc, SubnetSelection, Peer, Port } from '@aws-cdk/aws-ec2';
 import { SecureBucket } from '@cosmos-building-blocks/common';
-import { NPM_LOGIN, NPM_INSTALL } from './commands';
-import { BuildSpecBuilder } from './build-spec';
-import { BuildEnvironmentVariables, parseEnvs } from './utils';
+import { NPM_LOGIN, NPM_INSTALL } from '../commands';
+import { BuildSpecBuilder } from '../build-spec';
+import { BuildEnvironmentVariables, parseEnvs } from '../utils';
+import { SourceProvider, CodeCommitSourceProvider } from '../source';
 
 export interface CdkPipelineProps {
-  pipelineName?: string;
-  deployName?: string;
-  cdkRepo: IRepository;
-  cdkBranch?: string;
-  cdkWorkingDir?: string;
-  deployRole?: IRole;
-  deployVpc?: IVpc;
-  deploySubnets?: SubnetSelection;
-  deployEnvs?: BuildEnvironmentVariables;
-  deployStacks?: Array<Stack | string>;
-  deployDiffStage?: boolean;
-  npmKey?: string | BuildEnvironmentVariable;
+  readonly pipelineName?: string;
+  readonly deployName?: string;
+  readonly cdkSource?: SourceProvider;
+  readonly cdkRepo?: IRepository;
+  readonly cdkBranch?: string;
+  readonly cdkTrigger?: boolean;
+  readonly cdkWorkingDir?: string;
+  readonly deployRole?: IRole;
+  readonly deployVpc?: IVpc;
+  readonly deploySubnets?: SubnetSelection;
+  readonly deployEnvs?: BuildEnvironmentVariables;
+  readonly deployStacks?: Array<Stack | string>;
+  readonly deployDiffStage?: boolean;
+  readonly npmKey?: string | BuildEnvironmentVariable;
 }
 
 export interface AddDeployStackStageProps {
@@ -50,7 +47,7 @@ export interface AddDeployStackStageProps {
 
 export class CdkPipeline extends Construct {
   readonly stacks: IResolvable;
-  readonly cdkRepo: IRepository;
+  readonly cdkSource: SourceProvider;
   readonly deployRole: IRole;
   readonly deploy: Project;
   readonly pipeline: Pipeline;
@@ -62,8 +59,10 @@ export class CdkPipeline extends Construct {
     const {
       pipelineName,
       deployName,
+      cdkSource,
       cdkRepo,
-      cdkBranch = 'master',
+      cdkBranch,
+      cdkTrigger,
       cdkWorkingDir,
       deployRole,
       deployVpc,
@@ -82,11 +81,16 @@ export class CdkPipeline extends Construct {
       produce: () => stackNames(deployStacks || findAllStacksFromCdkApp(scope)),
     });
 
-    // Cross Stack Dependency issue since repo stack might differ from this stack
-    this.cdkRepo =
-      Stack.of(this) !== Stack.of(cdkRepo)
-        ? Repository.fromRepositoryName(this, cdkRepo.node.id, cdkRepo.repositoryName)
-        : cdkRepo;
+    const source = cdkRepo
+      ? ((new CodeCommitSourceProvider({
+          repo: cdkRepo,
+          branch: cdkBranch || 'master',
+          trigger: cdkTrigger || false,
+        }) as any) as SourceProvider)
+      : cdkSource;
+    if (!source) throw new Error('A source repository could not be found.');
+    this.cdkSource = source;
+    this.cdkSource.setup(this);
 
     this.deployRole =
       deployRole ||
@@ -146,13 +150,11 @@ export class CdkPipeline extends Construct {
       role: this.deployRole,
       vpc: deployVpc,
       subnetSelection: deploySubnets,
-      source: Source.codeCommit({
-        repository: this.cdkRepo,
-        branchOrRef: cdkBranch,
-      }),
+      source: this.cdkSource.source(),
       buildSpec: buildSpec,
       environment: {
-        buildImage: LinuxBuildImage.STANDARD_3_0,
+        computeType: ComputeType.SMALL,
+        buildImage: LinuxBuildImage.STANDARD_4_0,
         environmentVariables: parseEnvs(envs),
         privileged: true,
       },
@@ -178,16 +180,7 @@ export class CdkPipeline extends Construct {
 
     this.pipeline.addStage({
       stageName: 'Source',
-      actions: [
-        new CodeCommitSourceAction({
-          actionName: 'CdkCheckout',
-          role: this.pipeline.role,
-          repository: this.cdkRepo,
-          branch: cdkBranch,
-          output: sourceOutput,
-          trigger: CodeCommitTrigger.NONE,
-        }),
-      ],
+      actions: [this.cdkSource.sourceAction('CdkCheckout', this.pipeline.role, sourceOutput)],
     });
 
     if (deployDiffStage) {
@@ -226,7 +219,7 @@ export class CdkPipeline extends Construct {
   }
 
   addDeployStackStage(props: AddDeployStackStageProps): void {
-    const { name, stacks, pipeline = this.pipeline, envs = {}, isManualApprovalRequired = true, exclusive = true } =
+    const { name, stacks, pipeline = this.pipeline, envs = {}, isManualApprovalRequired = true, exclusive = false } =
       props || {};
 
     if (stacks && stacks.length === 0) {
@@ -237,13 +230,7 @@ export class CdkPipeline extends Construct {
     if (!cdkOutputArtifact) {
       cdkOutputArtifact = new Artifact('CdkOutput');
       pipeline.stages[0].addAction(
-        new CodeCommitSourceAction({
-          actionName: 'CdkCheckout',
-          role: pipeline.role,
-          repository: this.cdkRepo,
-          output: cdkOutputArtifact,
-          trigger: CodeCommitTrigger.NONE,
-        })
+        this.cdkSource.sourceAction('CdkCheckout', this.pipeline.role, cdkOutputArtifact, undefined, false)
       );
     }
 
@@ -275,50 +262,21 @@ export class CdkPipeline extends Construct {
   }
 
   preparePipeline(): void {
-    const allStacks = findAllStacksFromCdkApp(this);
     const stacks = Stack.of(this).resolve(this.stacks) as string[];
-    const stackStages = findStacksFromCdkDeployActions(this.pipeline.stages);
+    const stackStages = findStacksFromCdkDeployActions(this.pipeline);
     const remainingStacks = stacks.filter(x => !stackStages.includes(x));
-    const depStacks = stacks
-      .map(x => allStacks.find(y => x === y.stackName) as Stack)
-      .filter(x => x)
-      .filter(x => stackNames(findStackDependencies(x)).some(x => stackStages.includes(x)))
-      .map(x => x.stackName);
-    const beforeStacks = remainingStacks.filter(x => !depStacks.includes(x));
-    const afterStacks = remainingStacks.filter(x => depStacks.includes(x));
-
     const cdkOutputArtifact = findOutputFromCdkCheckout(this.pipeline.stages) as Artifact;
 
-    if (beforeStacks.length) {
+    if (remainingStacks.length) {
       this.pipeline.addStage({
         stageName: 'Deploy',
-        placement: {
-          justAfter: this.pipeline.stages[this.hasDiffStage ? 1 : 0],
-        },
         actions: [
           new CdkDeployAction({
             actionName: 'CdkDeploy',
             role: this.pipeline.role,
             project: this.deploy,
             input: cdkOutputArtifact,
-            stacks: beforeStacks,
-            hasDiffStage: this.hasDiffStage,
-            exclusive: true,
-          }),
-        ],
-      });
-    }
-
-    if (afterStacks.length) {
-      this.pipeline.addStage({
-        stageName: 'DeployDependantStacks',
-        actions: [
-          new CdkDeployAction({
-            actionName: 'DependantStackDeploy',
-            role: this.pipeline.role,
-            project: this.deploy,
-            input: cdkOutputArtifact,
-            stacks: afterStacks,
+            stacks: remainingStacks,
             hasDiffStage: this.hasDiffStage,
             exclusive: true,
           }),
@@ -328,10 +286,15 @@ export class CdkPipeline extends Construct {
   }
 }
 
+export interface CdkDeployActionProps extends CodeBuildActionProps {
+  stacks?: string[];
+  hasDiffStage?: boolean;
+  exclusive?: boolean;
+}
 export class CdkDeployAction extends CodeBuildAction {
-  // readonly actionProperties: ActionProperties & { environmentVariables?: BuildEnvironmentVariables };
-  stacks: string[];
-  constructor(props: CodeBuildActionProps & { stacks?: string[]; hasDiffStage?: boolean; exclusive?: boolean }) {
+  _props: CdkDeployActionProps;
+
+  constructor(props: CdkDeployActionProps) {
     const { stacks = [], hasDiffStage, environmentVariables, exclusive } = props;
 
     const envs: BuildEnvironmentVariables = {
@@ -358,21 +321,13 @@ export class CdkDeployAction extends CodeBuildAction {
       environmentVariables: parseEnvs(envs),
     });
 
-    this.stacks = stacks;
+    this._props = props;
   }
 
   static isCdkDeployAction(x: any): x is CdkDeployAction {
     return x instanceof CdkDeployAction;
   }
 }
-
-const findAllStacksFromCdkApp = (scope: IConstruct): Stack[] => {
-  const stacks = scope.node.scopes[0].node
-    .findAll()
-    .filter<Stack>(Stack.isStack)
-    .filter(x => !x.nested);
-  return stacks;
-};
 
 const findOutputFromCdkCheckout = (stages: IStage[]): Artifact | undefined => {
   const action = stages
@@ -386,20 +341,39 @@ const findOutputFromCdkCheckout = (stages: IStage[]): Artifact | undefined => {
   return action.actionProperties.outputs[0];
 };
 
-const findStacksFromCdkDeployActions = (stages: IStage[]): string[] =>
-  stages
+const findStacksFromCdkDeployActions = (pipeline: Pipeline): string[] =>
+  pipeline.stages
     .reduce<IAction[]>((res, item) => {
       res.push(...item.actions);
       return res;
     }, [])
     .filter(CdkDeployAction.isCdkDeployAction)
     .reduce<string[]>((res, item) => {
-      if (item.stacks) res.push(...item.stacks);
+      const stacks = item._props.stacks;
+      if (stacks) {
+        res.push(...stacks);
+        if (!item._props.exclusive) {
+          const allStacks = findAllStacksFromCdkApp(pipeline);
+          const deps = stacks
+            .map(x => allStacks.find(y => x === y.stackName) as Stack)
+            .filter(x => x)
+            .map(x => findStackDependencies(x))
+            .flat()
+            .map(x => x.stackName);
+          for (const stack of deps) {
+            if (!res.includes(stack)) res.push(stack);
+          }
+        }
+      }
       return res;
     }, []);
 
-const stackNames = (stacks: Array<Stack | string>): string[] => {
-  return stacks.map(x => (typeof x === 'string' ? x : x.stackName));
+const findAllStacksFromCdkApp = (scope: IConstruct): Stack[] => {
+  const stacks = scope.node.scopes[0].node
+    .findAll()
+    .filter<Stack>(Stack.isStack)
+    .filter(x => !x.nested);
+  return stacks;
 };
 
 const findStackDependencies = (stack: Stack): Stack[] => {
@@ -416,4 +390,8 @@ const findStackDependencies = (stack: Stack): Stack[] => {
     }, []);
 
   return deps;
+};
+
+const stackNames = (stacks: Array<Stack | string>): string[] => {
+  return stacks.map(x => (typeof x === 'string' ? x : x.stackName));
 };