+++ packages/@cosmos-building-blocks/pipeline/test/pipeline.test.ts
@@ -2,6 +2,9 @@ import { SynthUtils } from '@aws-cdk/assert';
 import { App, Stack } from '@aws-cdk/core';
 import { Repository } from '@aws-cdk/aws-codecommit';
 import { NodePipeline, DockerPipeline, CdkPipeline } from '../src';
+import { GithubEnterpriseConnection, GithubEnterpriseSourceProvider } from '../src/source';
+import { GithubEnterpriseHost } from '../src/source/github-enterprise-connection';
+import { Vpc } from '@aws-cdk/aws-ec2';
 
 describe('Node Pipeline', () => {
   const app = new App();
@@ -83,31 +86,35 @@ describe('Cdk Pipeline', () => {
     expect(cdk.pipeline.stages).toHaveLength(4);
     expect(synth.template).toMatchSnapshot();
   });
+});
 
-  test('cdk pipeline should support stack dependent deployments', () => {
-    const app = new App();
-    const stack = new Stack(app, 'Test');
-    const stack2 = new Stack(app, 'Test2');
+describe('Github Enterprise', () => {
+  const app = new App();
+  const stack = new Stack(app, 'Test');
 
-    const repo = new Repository(stack, 'Repo', {
-      repositoryName: 'repo',
-    });
-
-    const cdk = new CdkPipeline(stack, 'Pipeline', {
-      cdkRepo: repo,
-    });
-
-    cdk.addDeployStackStage({ name: 'TestStage', stacks: [stack2] });
-
-    // Should be able to add stacks after CdkPipeline
-    const stack3 = new Stack(app, 'Test3');
-    stack3.addDependency(stack2);
-    const stack4 = new Stack(app, 'Test4');
-    stack4.node.addDependency(stack3);
-
-    const synth = SynthUtils.synthesize(stack);
-
-    expect(cdk.pipeline.stages).toHaveLength(5);
-    expect(synth.template).toMatchSnapshot();
+  const repo = 'https://ghe.com/timpur/test.git';
+  const host = new GithubEnterpriseHost(stack, 'GHEHost', {
+    hostName: 'GHEHost',
+    vpc: Vpc.fromVpcAttributes(stack, 'Vpc', {
+      vpcId: '123',
+      availabilityZones: ['a'],
+      isolatedSubnetNames: ['App'],
+      isolatedSubnetIds: ['123'],
+      isolatedSubnetRouteTableIds: ['123'],
+    }),
+    subnets: [{ subnetGroupName: 'App' }],
+    endpoint: 'https://ghe.com/',
   });
+  const connection = new GithubEnterpriseConnection(stack, 'GHE', {
+    connectionName: 'GHE',
+    host: host,
+  });
+
+  const cdk = new CdkPipeline(stack, 'Pipeline', {
+    cdkSource: new GithubEnterpriseSourceProvider({ connection, repo, branch: 'master' }),
+  });
+
+  const synth = SynthUtils.synthesize(stack);
+
+  expect(synth.template).toMatchSnapshot();
 });